+++ integration/tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from helpers import context
+
+CTX = context.Context({'brig': 8082})
+
+@pytest.fixture
+def ctx(): return CTX
