+++ integration/tests/test_test.py
@@ -1,7 +1,5 @@
 from helpers import api
 
-ctx = api.Context({'brig': 8082})
-
-def test_brig_user():
-    with api.create_user(ctx) as r:
+def test_brig_user(ctx):
+    with ctx.create_user() as r:
         assert r.status_code == 201
